Move ToU64 to bitcoin
Outside of a single usage in p2p which can be replaced by a From call,
the ToU64 trait is only used in bitcoin.

Move ToU64 trait to bitcoin, keeping it pub for use in existing public
API code.
